Continued work on "select classes" functionality
As of now the Admin is able to select from a list of courses and edit
how many sections of the particular course they need.

Upon pressing a button, a list of all the sections (with section
requirements) is generated.

TO DO:

As of right now I still need to bind data from the Professor table to
the section info Professor comboBox control.

Create an INSERT query to insert all the section info into the database.