FIXED checkbox bug
Recently there was a slight bug with the checkboxes that were being
generated dynamically.

Turns out it was because of a variable that was being reset every time a
new courses' sections were being generated. A new variable has been
added to keep a consistent and unique count of ALL the sections (for ALL
courses) that are to be generated.

Long story short... PROBLEM SOLVED